[dev] 网络请求库开发

+++ app/src/main/java/com/pandaq/pandamvp/framework/AppBaseActivity.java
@@ -91,7 +91,7 @@ public abstract class AppBaseActivity<P extends BasePresenter> extends BaseActiv
                 .show();
     }
 
-    public void onError(int errCode, String errMsg) {
+    public void onError(long errCode, String errMsg) {
         Toaster.with(this)
                 .msg(errMsg)
                 .show();
